MISRA Rule 10.3 updates for MQTT files and iot_atomic_gcc.h (#724)
* Update MQTT files and iot_atomic_gcc.h for MISRA Rule 10.3 violations.

+++ ports/common/include/atomic/iot_atomic_gcc.h
@@ -49,8 +49,8 @@ static FORCE_INLINE uint32_t Atomic_CompareAndSwap_u32( uint32_t volatile * pDes
 {
     uint32_t swapped = 0;
 
-    /* This header file is used with only the gcc compiler which 
-     * requires an int parameter for this routine. */ 
+    /* This header file is used with only the gcc compiler which
+     * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
     if( __atomic_compare_exchange( pDestination,
                                    &comparand,
@@ -75,7 +75,7 @@ static FORCE_INLINE void * Atomic_Swap_Pointer( void * volatile * pDestination,
 {
     void * pOldValue = NULL;
 
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
     __atomic_exchange( pDestination, &pNewValue, &pOldValue, __ATOMIC_SEQ_CST );
@@ -115,10 +115,10 @@ static FORCE_INLINE uint32_t Atomic_CompareAndSwap_Pointer( void * volatile * pD
 static FORCE_INLINE uint32_t Atomic_Add_u32( uint32_t volatile * pAugend,
                                              uint32_t addend )
 {
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
-    return __atomic_fetch_add( pAugend, addend, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_add( pAugend, addend, __ATOMIC_SEQ_CST ) );
 }
 
 /*-----------------------------------------------------------*/
@@ -129,10 +129,10 @@ static FORCE_INLINE uint32_t Atomic_Add_u32( uint32_t volatile * pAugend,
 static FORCE_INLINE uint32_t Atomic_Subtract_u32( uint32_t volatile * pMinuend,
                                                   uint32_t subtrahend )
 {
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
-    return __atomic_fetch_sub( pMinuend, subtrahend, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_sub( pMinuend, subtrahend, __ATOMIC_SEQ_CST ) );
 }
 
 /*-----------------------------------------------------------*/
@@ -142,7 +142,7 @@ static FORCE_INLINE uint32_t Atomic_Subtract_u32( uint32_t volatile * pMinuend,
  */
 static FORCE_INLINE uint32_t Atomic_Increment_u32( uint32_t volatile * pAugend )
 {
-    return __atomic_fetch_add( pAugend, 1, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_add( pAugend, 1U, __ATOMIC_SEQ_CST ) );
 }
 
 /*-----------------------------------------------------------*/
@@ -152,7 +152,7 @@ static FORCE_INLINE uint32_t Atomic_Increment_u32( uint32_t volatile * pAugend )
  */
 static FORCE_INLINE uint32_t Atomic_Decrement_u32( uint32_t volatile * pMinuend )
 {
-    return __atomic_fetch_sub( pMinuend, 1, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_sub( pMinuend, 1U, __ATOMIC_SEQ_CST ) );
 }
 
 /*--------------------- Bitwise logic -------------------------*/
@@ -163,10 +163,10 @@ static FORCE_INLINE uint32_t Atomic_Decrement_u32( uint32_t volatile * pMinuend 
 static FORCE_INLINE uint32_t Atomic_OR_u32( uint32_t volatile * pOperand,
                                             uint32_t mask )
 {
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
-    return __atomic_fetch_or( pOperand, mask, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_or( pOperand, mask, __ATOMIC_SEQ_CST ) );
 }
 
 /*-----------------------------------------------------------*/
@@ -177,10 +177,10 @@ static FORCE_INLINE uint32_t Atomic_OR_u32( uint32_t volatile * pOperand,
 static FORCE_INLINE uint32_t Atomic_XOR_u32( uint32_t volatile * pOperand,
                                              uint32_t mask )
 {
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
-    return __atomic_fetch_xor( pOperand, mask, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_xor( pOperand, mask, __ATOMIC_SEQ_CST ) );
 }
 
 /*-----------------------------------------------------------*/
@@ -191,10 +191,10 @@ static FORCE_INLINE uint32_t Atomic_XOR_u32( uint32_t volatile * pOperand,
 static FORCE_INLINE uint32_t Atomic_AND_u32( uint32_t volatile * pOperand,
                                              uint32_t mask )
 {
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
-    return __atomic_fetch_and( pOperand, mask, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_and( pOperand, mask, __ATOMIC_SEQ_CST ) );
 }
 
 /*-----------------------------------------------------------*/
@@ -205,10 +205,10 @@ static FORCE_INLINE uint32_t Atomic_AND_u32( uint32_t volatile * pOperand,
 static FORCE_INLINE uint32_t Atomic_NAND_u32( uint32_t volatile * pOperand,
                                               uint32_t mask )
 {
-    /* This header file is used with only the gcc compiler which 
+    /* This header file is used with only the gcc compiler which
      * requires an int parameter for this routine. */
     /* coverity[misra_c_2012_directive_4_6_violation] */
-    return __atomic_fetch_nand( pOperand, mask, __ATOMIC_SEQ_CST );
+    return ( uint32_t ) ( __atomic_fetch_nand( pOperand, mask, __ATOMIC_SEQ_CST ) );
 }
 
 #endif /* IOT_ATOMIC_GCC_H_ */
